pyproject.toml
[tool.poetry]
name = "dominant-colors"
version = "0.1.0"
description = ""
authors = ["olast45"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
matplotlib = "^3.8.4"
scikit-learn = "^1.4.2"
scipy = "^1.13.0"
pandas = "^2.2.2"
numpy = "^1.26.4"
streamlit = "^1.33.0"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
